Changing LICENSE from Apache 2.0 to CC BY 4.0 International
In order to be included in the "Awesomeverse", the license must be
a CC license. Creative Commons Attribution 4.0 International is the
nearest match to the Apache License 2.0 used previously.

The only major difference is the patent grant clause.
However, this repository is only documentational, hence no code,
nor patents thereof used.

Part of #9.